Perform APES model selection

Perform APES model selection

apes(
  model,
  k = NULL,
  estimator = "leaps",
  time_limit = 10L,
  verbose = FALSE,
  n_boot = 0L,
  workers = 1L
)

Arguments

model
A "glm" class object

k
Model size to explore. Default to NULL, which searches through all variables in the model. Alternatively, user can input a vector. If estimator is:

"leaps", then models up to size max(k) wil be explored.

"mio", then models with specified values will be explored.

estimator
Either "leaps" (default) or "mio", which correspond to optimisation algorithms available in the leaps and bestsubset package, respectively.

time_limit
The time limit for the maximum time allocated to each model size model when the "mio" estimator was selected. It will not affect the speed if leaps.

verbose
Whether to print off messages during computations

n_boot
Number of bootstrap runs, default to 0, which doesn't perform any sampling.

workers
Number of cores used for parallel processing for the bootstrap

Examples