Perform APES-logistic regression using either the leaps or bestsubset package

Perform APES-logistic regression using either the leaps or bestsubset package

apes_logit(x, y, Pi, k, estimator = "leaps", time.limit = 60)

Arguments

x
The n by p design matrix for logistic regression

y
The response vector for logistic regression of length n

Pi
Esimated probabilities of each observation using a baseline model. It should be a vector of length n.

k
Model size to explore. If leaps was selected as the estimator, then model up to size max(k) will be explored. If mio was selected as the estimator, then model sizes specified in k will be explored.

estimator
Either "leaps" or "mio", which correspond to optimisation algorithms available in the leaps and bestsubset package, respectively.

time.limit
The time limit for the maximum time allocated to each model size model when the "mio" estimator was selected. It will not affect the speed if leaps

Value

A list of results:

apesModelDf: A data.frame (tibble) containing the optimal model of each size and associated statistics

apesMleBeta: A matrix of MLE for the most optimal model of each model size.

apesMleBetaBinary: A data.frame (tibble), which has a column indicating if a variable has been selected into the most optimal model of each size. It is a binarised, linear version of the previous apesMleBeta output.

apesTimeDiff: A difftime object of the computational time.

selectedModelBeta: A matrix containing the MLE correspond to the AIC and BIC selected model. A subset of the apesMleBeta output.

aicWeights: A matrix of weights corresponding to the AIC_j/sum(AIC_j), where AIC_j denotes the most optimal model of the size j in the apesModelDf output.

bicWeights: A matrix of weights corresponding to the BIC_j/sum(BIC_j), where BIC_j denotes the most optimal model of the size j in the apesModelDf output.

modelAvgBeta: A matrix of AIC and BIC weighted coefficients. It is identical to the matrix product of apesMleBeta and aicWeights (or bicWeights).

responseTibble: A data.frame (tibble) of estimates for each observations, e.g. initial fitted response, final fitted response.

Examples

set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x %*% beta))
Pi = glm.fit(x = x, y = y, family = binomial(link = "logit"))$fitted.values
apesLeapsResult = apes_logit(x = x, y = y, Pi = Pi, k = k,
                               estimator = "leaps")

#> [1] "Finished solving linear regression approximation"

apesLeapsResult$apesModelDf

#> # A tibble: 10 x 8
#>    method modelName modelSize apesMleLoglike mleAIC mleBIC status
#>    <chr>  <chr>         <dbl>          <dbl>  <dbl>  <dbl> <chr> 
#>  1 apes   apesMode…         2          -59.4   123.   128. leaps…
#>  2 apes   apesMode…         3          -53.3   113.   120. leaps…
#>  3 apes   apesMode…         4          -51.8   112.   122. leaps…
#>  4 apes   apesMode…         5          -51.2   112.   125. leaps…
#>  5 apes   apesMode…         6          -50.6   113.   129. leaps…
#>  6 apes   apesMode…         7          -50.1   114.   132. leaps…
#>  7 apes   apesMode…         8          -50.1   116.   137. leaps…
#>  8 apes   apesMode…         9          -50.1   118.   142. leaps…
#>  9 apes   apesMode…        10          -50.0   120.   146. leaps…
#> 10 apes   apesMode…        11          -49.9   122.   151. leaps…
#> # … with 1 more variable: icOptimalModels <chr>

# NOT RUN {
## You need to install Gurobi before uing the mio option
apesMioResult = apes_logit(x = x, y = y, Pi = Pi, k = k,
                               estimator = "mio", time.limit = 5)
# }