Bootstrap version of apes_logit

Combine apes_logit results

apes_logit_mio_combine(listApes)

Arguments

listApes
a list of APES output with the MIO optimisation algorithm

Examples

set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)
y = rbinom(n = n, size = 1, prob = expit(x %*% beta))
Pi = glm.fit(x = x, y = y, family = binomial(link = "logit"))$fitted.values
firstApes = apes_logit(x = x, y = y, Pi = Pi, k = 1, estimator = "mio")

#> 0. Computing max eigenvalue of X^T X, for the step size in projected gradient descent.
#> 1. Solving best subset selection with k=1.
#>   a. Performing projected gradient runs: 1 ... 10 ... 20 ... 30 ... 40 ... 50 ... 
#>   b. Running Gurobi's mixed integer program solver ... Return status: OPTIMAL.

secondApes = apes_logit(x = x, y = y, Pi = Pi, k = 2, estimator = "mio")

#> 0. Computing max eigenvalue of X^T X, for the step size in projected gradient descent.
#> 1. Solving best subset selection with k=2.
#>   a. Performing projected gradient runs: 1 ... 10 ... 20 ... 30 ... 40 ... 50 ... 
#>   b. Running Gurobi's mixed integer program solver ... Return status: OPTIMAL.

listApes = list(firstApes, secondApes)
finalResult = apes_logit_mio_combine(listApes)

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

print(finalResult)

#> $apesModelDf
#> # A tibble: 2 x 9
#>   listName method modelName modelSize apesMleLoglike mleAIC mleBIC status
#>   <chr>    <chr>  <chr>         <dbl>          <dbl>  <dbl>  <dbl> <chr> 
#> 1 1        apes   apesMode~         2          -59.4   123.   128. OPTIM~
#> 2 2        apes   apesMode~         3          -53.3   113.   120. OPTIM~
#> # ... with 1 more variable: icOptimalModels <chr>
#> 
#> $apesMleBeta
#>     apesModel_2 apesModel_3
#> Int  -0.1837717  -0.0541247
#> X1    0.0000000   0.9097537
#> X2   -1.0242090  -1.1231560
#> X3    0.0000000   0.0000000
#> X4    0.0000000   0.0000000
#> X5    0.0000000   0.0000000
#> X6    0.0000000   0.0000000
#> X7    0.0000000   0.0000000
#> X8    0.0000000   0.0000000
#> X9    0.0000000   0.0000000
#> X10   0.0000000   0.0000000
#> 
#> $apesMleBetaBinary
#> # A tibble: 22 x 4
#>    listName variables modelName   fittedBeta
#>    <chr>    <fct>     <chr>       <lgl>     
#>  1 1        Int       apesModel_2 TRUE      
#>  2 1        X1        apesModel_2 FALSE     
#>  3 1        X2        apesModel_2 TRUE      
#>  4 1        X3        apesModel_2 FALSE     
#>  5 1        X4        apesModel_2 FALSE     
#>  6 1        X5        apesModel_2 FALSE     
#>  7 1        X6        apesModel_2 FALSE     
#>  8 1        X7        apesModel_2 FALSE     
#>  9 1        X8        apesModel_2 FALSE     
#> 10 1        X9        apesModel_2 FALSE     
#> # ... with 12 more rows
#> 
#> $apesTimeDiff
#> Time differences in mins
#>           1           2 
#> 0.002363968 0.003840450 
#> 
#> $selectedModelBeta
#>           [,1]       [,2]
#> Int -0.0541247 -0.0541247
#> X1   0.9097537  0.9097537
#> X2  -1.1231560 -1.1231560
#> X3   0.0000000  0.0000000
#> X4   0.0000000  0.0000000
#> X5   0.0000000  0.0000000
#> X6   0.0000000  0.0000000
#> X7   0.0000000  0.0000000
#> X8   0.0000000  0.0000000
#> X9   0.0000000  0.0000000
#> X10  0.0000000  0.0000000
#> 
#> $responseTibble
#> # A tibble: 100 x 6
#>    obsNum     y     Pi linearY apesMinAicProb apesMinBicProb
#>    <chr>  <int>  <dbl>   <dbl>          <dbl>          <dbl>
#>  1 obs1       0 0.824    -4.13         0.694          0.694 
#>  2 obs2       0 0.290    -2.30         0.333          0.333 
#>  3 obs3       1 0.396     2.10         0.467          0.467 
#>  4 obs4       1 0.237     3.05         0.198          0.198 
#>  5 obs5       1 0.523     2.00         0.716          0.716 
#>  6 obs6       0 0.339    -2.18         0.418          0.418 
#>  7 obs7       0 0.114    -3.18         0.131          0.131 
#>  8 obs8       1 0.878     3.11         0.814          0.814 
#>  9 obs9       0 0.0111   -5.50         0.0636         0.0636
#> 10 obs10      0 0.0401   -4.22         0.174          0.174 
#> # ... with 90 more rows
#> 
#> $modelAvgBeta
#>     aicWeightCoef bicWeightCoef
#> Int   -0.05490941   -0.05696541
#> X1     0.90424733    0.88982001
#> X2    -1.12255716   -1.12098801
#> X3     0.00000000    0.00000000
#> X4     0.00000000    0.00000000
#> X5     0.00000000    0.00000000
#> X6     0.00000000    0.00000000
#> X7     0.00000000    0.00000000
#> X8     0.00000000    0.00000000
#> X9     0.00000000    0.00000000
#> X10    0.00000000    0.00000000
#> 
#> $aicWeights
#>             [,1]
#> [1,] 0.006052629
#> [2,] 0.993947371
#> 
#> $bicWeights
#>            [,1]
#> [1,] 0.02191112
#> [2,] 0.97808888
#>