Perform APES model selection

Approximated exhaustive selection will be performed on the input GLM model.

apes(
  model,
  k = NULL,
  estimator = "leaps",
  time_limit = 10L,
  really_big = FALSE,
  verbose = FALSE,
  n_boot = 0,
  workers = 1L
)

Arguments

model
A "full" model with all variables of interest fitted to it. Accepts either a "glm" class or "coxph" class object.

k
Model size to explore. Default to NULL, which searches through all variables in the model. Alternatively, user can input a vector. If estimator is:

"leaps", then models up to size max(k) wil be explored.

"mio", then models with specified values will be explored.

estimator
Either "leaps" (default) or "mio", which correspond to optimisation algorithms available in the leaps and bestsubset package, respectively.

time_limit
The time limit for the maximum time allocated to each model size model when the "mio" estimator was selected. It will not affect the computational speed if "leaps" is selected as the estimator.

really_big
If set to FALSE (by default), then it will prevent variable selection greater than 30 variables. We recommend only setting this argument to TRUE if the number of variables exceed 30 while simultaneous setting the "estimator" argument to "mio".

verbose
Whether to print off messages during computations.

n_boot
Number of bootstrap runs, default to 0, which doesn't perform any sampling.

workers
Number of cores used for parallel processing for the bootstrap

Value

Either an object of class "apes" in case "n_boot" is set to zero or an object of class "boot_apes" in case "n_boot" is set to a positive integer.

Details

This is the main function of the APES package. Note that the feature dimensions equals to the number of columns if all columns are numeric. If factor/categorical variables are present in the input "glm" object, then the feature dimension also includes all the levels of factors. See vignette on birth weight.

Examples

set.seed(10)
n = 100
p = 10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)

## Logistic regression example
y = rbinom(n = n, size = 1, prob = expit(x %*% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "binomial")
apes_result = apes(model = model)
apes_result

#> Time taken:  0.0006356676  minutes 
#> 
#>  APES - AIC selected the following variables 
#> intercept        X1        X2        X3        X4        X5        X6        X7 
#>    -0.041     0.912    -1.145     0.000     0.000     0.000    -0.392     0.000 
#>        X8        X9       X10 
#>     0.000     0.000     0.000 
#> 
#>  APES - BIC selected the following variables 
#> intercept        X1        X2        X3        X4        X5        X6        X7 
#>    -0.054     0.910    -1.123     0.000     0.000     0.000     0.000     0.000 
#>        X8        X9       X10 
#>     0.000     0.000     0.000 

class(apes_result)

#> [1] "apes"

names(apes_result)

#> [1] "apes_model_df"        "apes_mle_beta"        "apes_mle_beta_binary"
#> [4] "time_used"            "selected_model_beta"  "model_avg_beta"      
#> [7] "response_tibble"     

## Poisson regression example
y = rpois(n = n, lambda = exp(x %*% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "poisson")
apes(model = model)

#> Time taken:  8.066495e-06  minutes 
#> 
#>  APES - AIC selected the following variables 
#> intercept        X1        X2        X3        X4        X5        X6        X7 
#>    -0.216     0.935    -1.133     0.000     0.009     0.000     0.022     0.000 
#>        X8        X9       X10 
#>    -0.054     0.000     0.000 
#> 
#>  APES - BIC selected the following variables 
#> intercept        X1        X2        X3        X4        X5        X6        X7 
#>    -0.216     0.935    -1.133     0.000     0.009     0.000     0.022     0.000 
#>        X8        X9       X10 
#>    -0.054     0.000     0.000 

## Bootstrap examples
apes(model = model, n_boot = 2)

#> Time taken:  0.003102752  minutes 
#> Total number of bootstrap APES results:  2

apes(model = model, n_boot = 2, workers = 1)

#> Time taken:  0.003590167  minutes 
#> Total number of bootstrap APES results:  2

## apes(model = model, n_boot = 2, workers = 2)

## Cox regression example
hx = exp(x %*% beta)
time = rexp(n, hx)
censor = rbinom(n = n, prob = 0.3, size = 1) # censoring indicator
data = data.frame(x, time = time, censor = censor)
model = survival::coxph(survival::Surv(time, censor) ~ ., data = data)
apes(model = model)

#> Time taken:  6.850561e-06  minutes 
#> 
#>  APES - AIC selected the following variables 
#> intercept        X1        X2        X3        X4        X5        X6        X7 
#>     0.000     0.874    -1.172     0.000     0.387     0.000     0.000     0.000 
#>        X8        X9       X10 
#>     0.000    -0.395     0.000 
#> 
#>  APES - BIC selected the following variables 
#> intercept        X1        X2        X3        X4        X5        X6        X7 
#>     0.000     0.777    -1.082     0.000     0.000     0.000     0.000     0.000 
#>        X8        X9       X10 
#>     0.000     0.000     0.000