Perform APES model selection

Perform APES model selection

apes(
  model,
  k = NULL,
  estimator = "leaps",
  time_limit = 10L,
  verbose = FALSE,
  n_boot = 0L,
  workers = 1L
)

Arguments

model
A "glm" class object

k
Model size to explore. Default to NULL, which searches through all variables in the model. Alternatively, user can input a vector. If estimator is:

"leaps", then models up to size max(k) wil be explored.

"mio", then models with specified values will be explored.

estimator
Either "leaps" (default) or "mio", which correspond to optimisation algorithms available in the leaps and bestsubset package, respectively.

time_limit
The time limit for the maximum time allocated to each model size model when the "mio" estimator was selected. It will not affect the speed if leaps.

verbose
Whether to print off messages during computations

n_boot
Number of bootstrap runs, default to 0, which doesn't perform any sampling.

workers
Number of cores used for parallel processing for the bootstrap

Examples

set.seed(10)
n = 100
p = 10
k = 1:10
beta = c(1, -1, rep(0, p-2))
x = matrix(rnorm(n*p), ncol = p)
colnames(x) = paste0("X", 1:p)

## Logistic regression example
y = rbinom(n = n, size = 1, prob = expit(x %*% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "binomial")
apes(model = model)

#> Time taken: 
#> Time difference of 2.358357e-05 mins
#> 
#>  APES model selection data frame: 
#> # A tibble: 10 x 8
#>    method model_name model_size apes_mle_loglike mle_aic mle_bic status
#>    <chr>  <chr>           <dbl>            <dbl>   <dbl>   <dbl> <chr> 
#>  1 apes   apes_mode…          2            -59.4    123.    128. leaps…
#>  2 apes   apes_mode…          3            -53.3    113.    120. leaps…
#>  3 apes   apes_mode…          4            -51.8    112.    122. leaps…
#>  4 apes   apes_mode…          5            -51.2    112.    125. leaps…
#>  5 apes   apes_mode…          6            -50.6    113.    129. leaps…
#>  6 apes   apes_mode…          7            -50.1    114.    132. leaps…
#>  7 apes   apes_mode…          8            -50.1    116.    137. leaps…
#>  8 apes   apes_mode…          9            -50.1    118.    142. leaps…
#>  9 apes   apes_mode…         10            -50.0    120.    146. leaps…
#> 10 apes   apes_mode…         11            -49.9    122.    151. leaps…
#> # … with 1 more variable: ic_opt_models <chr>

## Poisson regression example
y = rpois(n = n, lambda = exp(x %*% beta))
data = data.frame(y, x)
model = glm(y ~ ., data = data, family = "poisson")
apes(model = model)

#> Time taken: 
#> Time difference of 9.000301e-06 mins
#> 
#>  APES model selection data frame: 
#> # A tibble: 10 x 8
#>    method model_name model_size apes_mle_loglike mle_aic mle_bic status
#>    <chr>  <chr>           <dbl>            <dbl>   <dbl>   <dbl> <chr> 
#>  1 apes   apes_mode…          2            -206.    416.    422. leaps…
#>  2 apes   apes_mode…          3            -206.    418.    426. leaps…
#>  3 apes   apes_mode…          4            -206.    420.    430. leaps…
#>  4 apes   apes_mode…          5            -203.    416.    429. leaps…
#>  5 apes   apes_mode…          6            -132.    277.    292. leaps…
#>  6 apes   apes_mode…          7            -132.    279.    297. leaps…
#>  7 apes   apes_mode…          8            -132.    280.    301. leaps…
#>  8 apes   apes_mode…          9            -131.    281.    304. leaps…
#>  9 apes   apes_mode…         10            -131.    282.    308. leaps…
#> 10 apes   apes_mode…         11            -131.    284.    313. leaps…
#> # … with 1 more variable: ic_opt_models <chr>

## Bootstrap examples
apes(model = model, n_boot = 2)

#> Time taken:  0.003118002  minutes 
#> Total number of bootstrap APES results:  2

apes(model = model, n_boot = 2, workers = 1)

#> Time taken:  0.001481799  minutes 
#> Total number of bootstrap APES results:  2

apes(model = model, n_boot = 2, workers = 2)

#> Time taken:  0.09137141  minutes 
#> Total number of bootstrap APES results:  2

## Cox regression example
hx = exp(x %*% beta)
time = rexp(n, hx)
censor = rbinom(n = n, prob = 0.3, size = 1) # censoring indicator
data = data.frame(x, time = time, censor = censor)
model = survival::coxph(survival::Surv(time, censor) ~ ., data = data)
apes(model = model)

#> Time taken: 
#> Time difference of 1.078049e-05 mins
#> 
#>  APES model selection data frame: 
#> # A tibble: 10 x 8
#>    method model_name model_size apes_mle_loglike mle_aic mle_bic status
#>    <chr>  <chr>           <dbl>            <dbl>   <dbl>   <dbl> <chr> 
#>  1 apes   apes_mode…          2            -74.7    153.    159. leaps…
#>  2 apes   apes_mode…          3            -67.3    141.    148. leaps…
#>  3 apes   apes_mode…          4            -65.6    139.    150. leaps…
#>  4 apes   apes_mode…          5            -64.7    139.    152. leaps…
#>  5 apes   apes_mode…          6            -64.2    140.    156. leaps…
#>  6 apes   apes_mode…          7            -64.2    142.    161. leaps…
#>  7 apes   apes_mode…          8            -64.1    144.    165. leaps…
#>  8 apes   apes_mode…          9            -64.1    146.    170. leaps…
#>  9 apes   apes_mode…         10            -64.1    148.    174. leaps…
#> 10 apes   apes_mode…         11            -64.1    150.    179. leaps…
#> # … with 1 more variable: ic_opt_models <chr>